requirements.txt
# Core ML/Data Science Libraries
tensorflow==2.15.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.4
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database Connections
pymongo==4.6.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Data Sources
yfinance==0.2.28
alpha-vantage==2.3.1
requests==2.31.0

# Statistical Analysis
statsmodels==0.14.1
scipy==1.11.4

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
python-multipart==0.0.6

# Development Tools
pytest==7.4.3
black==23.12.0
flake8==6.1.0
pre-commit==3.6.0

# Additional ML Libraries
xgboost==2.0.2
lightgbm==4.1.0

# Authentication & Security
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Real-time & WebSocket
websockets==12.0
fastapi-websocket==0.1.5

# Additional utilities
psutil==5.9.6
